reference/scenarios/azure-openai/pyproject.toml
[project]
name = "azure-openai-reference-test"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
    "openai==2.31.0",
    "genai-reference-shared",
]

[tool.uv.sources]
genai-reference-shared = { path = "../../shared", editable = true }

[tool.uv]
package = false
